requirements.txt
pandas>=2.2.2
sqlalchemy>=2.0.30
pymysql>=1.1.0
python-dotenv>=1.0.1

numpy>=1.26.4
scikit-learn>=1.5.0
mlxtend>=0.23.1

matplotlib>=3.9.0
seaborn>=0.13.2
plotly>=5.22.0

streamlit>=1.35.0

jupyter>=1.0.0
ipykernel>=6.29.0